refactor: 페이지명 오류 수정
